LayoutTests/platform/blackberry/accessibility/aria-menubar-menuitems-expected.txt

Menu 1
Menu item 1.1
Menu 2
Menu item 2.1
Menu item 2.2
This tests that ARIA menu items are properly exposed as children for menu bars

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



FAIL menuBar.childrenCount should be 2. Was 0.

FAIL menuItem.childrenCount should be 2. Was 0.

PASS menuSubItem1.childrenCount is 0

FAIL menuSubItem2.childrenCount should be 1. Was 0.

PASS subMenuSubItem.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
